+++ pkg/oci/mock.go
@@ -21,7 +21,7 @@ func (m *MockClient) Verify(ctx context.Context) error {
 	return nil
 }
 
-func (m *MockClient) Subscribe(ctx context.Context) (<-chan Image, <-chan error) {
+func (m *MockClient) Subscribe(ctx context.Context) (<-chan ImageEvent, <-chan error) {
 	return nil, nil
 }