+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#331](https://github.com/XenitAB/spegel/pull/331) Document possible modifications required for k8s-digester.
 - [#337](https://github.com/XenitAB/spegel/pull/337) Add HTTP bootstrapper.
 - [#340](https://github.com/XenitAB/spegel/pull/340) Add Talos to compatibility.
+- [#343](https://github.com/XenitAB/spegel/pull/343) Implement image event and add support for delete events.
 
 ### Changed